SIM-2800: Move demos into the fable_saga package itself for easier access.
