+++ fable_saga/demos/space_colony/resources/interactable_objects.yaml
