+++ fable_saga/demos/space_colony/resources/locations.yaml
