+++ fable_saga/demos/space_colony/resources/skills.yaml
